requirements.txt
numpy
pandas
matplotlib
scikit-learn
torch
torchvision
Pillow
albumentations